Modify delete to await user confirmation
Delete instantly removes contact if input is valid. To improve
user experience, request for confirmation before deleting.

Let's,
* Introduce state through a State interface and StateManager class
* Create a ConfirmationPendingResult class to represent
result returns of operations that require confirmation
* Create a ConfirmCommand
* Modify LogicManager to create ConfirmCommands instead of
normally parsing inputs through AddressBookParser if the app is
awaiting user confirmation.

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -25,6 +25,8 @@ public class DeleteCommand extends Command {
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
 
+    public static final String MESSAGE_DELETE_PERSON_CONFIRM = "Confirm deletion [y/n] of:\n%1$s?";
+
     private final Index targetIndex;
 
     public DeleteCommand(Index targetIndex) {
@@ -41,8 +43,17 @@ public class DeleteCommand extends Command {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deletePerson(personToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
+        return new ConfirmationPendingResult(
+                String.format(MESSAGE_DELETE_PERSON_CONFIRM, Messages.format(personToDelete)),
+                false,
+                false,
+                () -> {
+                    model.deletePerson(personToDelete);
+                    return new CommandResult(
+                            String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete))
+                    );
+                }
+        );
     }
 
     @Override
